Prepare 1.0.1 release, update to Kotlin 1.5.0

+++ codegen/src/main/kotlin/com/kobil/vertx/ebservice/codegen/ServiceProcessor.kt
@@ -105,7 +105,7 @@ class ServiceProcessor : AbstractProcessor() {
         .toSet()
 
       ServiceFun(
-        kmFunction.name, returnType, parameters, overloadId,
+        kmFunction.name, returnType, parameters, overloadId.toUInt(),
         isSuspend = kmFunction.isSuspend
       )
     }
